requirements.txt
django
pymysql
hdfs
thrift
happybase
pyjwt
pillow